Bring up interface when adding it to bridge
When deploying on metal with MAAS enlisted and bound interfaces
will be brought up at boot automatically.

However, the user may choose to use interfaces that are either hot
plugged or not configured/enlisted with substrate.

Thus the charm should bring up any interface before adding it to
a bridge.

Change-Id: Ia27fd7f2a823cac780d72e08f8c3fc54c7f9401c