Remove dependency on openvswitch-switch daemons
While the ovn packages have a dependency on the
`openvswitch-switch` and `openvswitch-common` packages the OVN
central components have no need for having the daemons running.

Also fixes an issue with the ovn-central systemd service file
which lead to unwanted reverse dependency between `ovn-northd` and
the database services.  I.e. we do not want a restart of
`ovn-northd` to lead to a reastart of the `ovn-nb-ovsdb` service.

Remove loading of `openvswitch` kernel module from LXD profile,
add in iptables instead as we will need this to protect the SB
DB listener created for accepting connections from `ovn-northd`

Change-Id: I134ec560c5522f0c657c203a4cd828e0e5c6f3cb