Move chassis code to layer-ovn
The two charms ``ovn-chassis`` and ``ovn-dedicated-chassis`` are apart
from metadata identical.  Move the shared code to layer.

Add missing LXD profile.

Reasoning behind updates to functional test bundles:
The `ovn-central` charm has been changed to disable openvswitch
components, and consequently it is no longer suitable as being a
principle charm for testing the ``ovn-chassis`` subordinate charm.

I chose to use the ``magpie`` charm as a principle charm for the
test as we regularily use it in our teams testing already and it
is most likely to stay on the bleeding edge wrt. series support.

Change-Id: If58c870481876eeac26127ca2459d34b6617cb6e