Make ovsdb-server inactivity_probe configurable
By default Open vSwitch ovsdb-server sends a inactivity probe
on client connections every 5 seconds.  If no response is received
by the next time a probe would be sent the connection is termiated.

Depending on type of CMS and load the default of 5 seconds may not
be sufficient. Consequences of a dropped connection can be increased
load due to recalculation and can also interrupt in-flight requests
from the control plane.

The OpenStack Neutron OVN driver currently assumes a 60s probe
interval, so we align the charm default with that.

Change-Id: I7918dfd0b5165ff20a65eeaca6b4934e2b1b9697