Actions: Move concrete charm module to works with layer-openstack
This patchset fixes the reference bug by moving the concrete charm class
module to lib/charm/openstack/ from lib/charms.  This is so that the
bus.discover() function in layer-openstack [1] can find the concrete
class and instantiate it.  The patchset is dependent on the layer-ovn
(linked below).

[1]: https://review.opendev.org/p/openstack/charm-layer-openstack
Depends On: https://github.com/openstack-charmers/charm-layer-ovn/pull/4

Change-Id: If9ef4699021f36a5e372c57056c89ded499403a3
Closes-Bug: #1861069