Fix handling of linux bond interfaces
At present linux bonds will be ignored when mentioned directly
in configuration options parsed by the ``BridgePortInterface``
data structure.

If a mac address of a physical interface making up the bond is
listed the code correctly finds the bond name and adds that to
the bridge.

Virtual interfaces are ignored when then map is built, linux bonds
are virtual interfaces.

This patch fixes the problem by checking physical interfaces for
bond membership while building the map. If an interface is member
of a bond add the bond name to the map.

Previously the bond membership check was only done when looking
up interface name from mac after building the map.

Change-Id: I941609236ff2bad62b191d0adf4a36bd2b34a42b
Closes-Bug: #1883244