Do not render cluster remote address on charm leader
The `ovn-ctl` script uses the combination of presence of
`--db-*-cluster-local-addr` and non-presence of
`--db-*-cluster-remote-addr` to decide when to do certain actions. [0]

Update the charm to not render the `--db-*-cluster-remote-addr`
argument on the charm leader.

Note that once the clustered database file has been created and
peers have joined the OVSDB server processes uses the contents
of the database file to connect to their peers and not the
information provided as arguments to the `ovn-ctl` script.

0: https://github.com/ovn-org/ovn/commit/67e2f386cc838d0b0f9b4b5da7fe611e1113b70c

Change-Id: I88b4aba31c8c05dba6f3f4ade5a8951613483d64
Closes-Bug: #1907081