Add xena bundles
- add non-voting impish
- drop groovy bundle
- rebuild to pick up charm-helpers changes
- update tox/pip.sh to ensure setuptools<50.0.0

Note that this charm doesn't require interim OpenStack releases (e.g.
focal-ussuri, -xena, -victoria) as it isn't based on OpenStack releases.
The ovn-chassis charm doesn't have a source or openstack-origin.  The
bionic-train and bionic-ussuri jobs are retained for historical
perspectives as they specify a PPA for the magpie charm in those
bundles.

Change-Id: I9840db1b88781d074dfbce8db875f458e5fd090a