Fix UCA source in bundles
At present the bundles will not test what is on the tin.

The ovn-chassis charm is a subordinate charm and requires its
principle charm to configure the correct UCA pocket on the
unit.  The magpie charm has grown a ``source`` configuration
option for this purpose.

This patch updates bundles to use the correct version of the
magpie charm and adds the source configuration option to its
bundle configuration.

Change-Id: I929043f97cb944e17da350aa476cd38902472557