Default to OVN 22.03 for fresh deployments on Focal
Add `ovn-source` configuration option which should be used in
addition to the `source` configuration option.

The default of the `ovn-source` configuration option is
determined at runtime based on whether the end user has provided
config, whether its a fresh or upgraded charm deployment and which
series it is deployed on.

Change-Id: I54f2339a16bd958c29f877394209039f345e106e
(cherry picked from commit 855bd843b532acdfecf176608bce2bffe3b29367)