Add bindep for CI Python dependencies and fix tox.ini
The opendev CI environment has recently transitioned to using
Ubuntu Jammy as the base OS for test runs.

A side effect of this transition is a change in some of the
dependencies for lint and unit tests.

tox.ini:
* Remove pins.
* Properly format passenv list.
* Drop the outer tox.ini func-target hack.

Drop the incepted tox in src/test-requirements.txt.

Drop pip.sh script.

Change-Id: I9d52fa71ce98f6e77462d6f8ab52b62fd787c356