test: switch principal charm from magpie to ubuntu
Since charm-ovn-chassis is a subordinate charm, it needs a principal
charm to deploy to. This used to be 'magpie' but that charm
doesn't have a 'noble' series release. To allow us to test ovn-chassis
noble release, we are switching the principal charm to 'ubuntu'.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1308
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>