CI: Add github actions
Main goal of this commit is to introduce Github CI action. These
include:
  * Commit compliance checks
  * Unit tests
  * Linters
  * Charm build
  * Functional tests

As a result of moving CI to github, which uses LXD as a juju
substrate, the gate bundle (jammy-caracal) has additional constraints
that force juju to deploy ovn-dedicated-chassis on a VM instead of a
container.

In addition, zaza dependencies for functional tests were upgraded to
Juju3, because Juju2.9 can't properly apply "virtual-machine"
constraint on LXD.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>