Ensure juju displays errmsg from SSL issues
There are times when the certificate expires, and the charm goes into error
state. This change ensures the right message is displayed, and the right actions
can be taken. In errored state the vault reissue action doesn't work, this
update would allow for the certs to be updated as the charm is not in error
state.

Signed-off-by: Arif Ali <arif.ali@canonical.com>
(cherry picked from commit 37465f494b3c224af1776435058439106eb0bc5d)

Conflicts:
	Commit message updated to omit func-test-pr tag
	ovn-central/src/tests/tests.yaml
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>